ticket-booking-app/backend/requirements.txt
fastapi
passlib[bcrypt]
uvicorn[standard]
python-jose[cryptography]
sqlalchemy
psycopg2-binary
alembic
redis
celery
python-dotenv
pydantic
email-validator